+++ test.sh
@@ -1,1 +1,1 @@
-g++ -o test.prg solver_mm.cpp interpol.cpp string_handler.cpp -I$EIGEN3_INCLUDE_DIR
+g++ -std=c++11 -o test.prg solver_mm.cpp interpol.cpp string_handler.cpp derivatives_handler.cpp -I$EIGEN3_INCLUDE_DIR -L/usr/local/lib -I/usr/local/include